Clean the sllw/slliw/sraw/sraiw/srlw/srliw
Use the sll to instead the sllw,
use the slli to insted the slliw,
use the sra to insated the sraw,
use the srai to instead the sraiw,
use the srl to instead the srlw,
use the srli to instead the srliw.

After this patch, the RV32 code will be cleaned in
the sllw/slliw/sraw/sraiw/srlw/srliw.