Modify the RISCV32 name system and improve the thread code
Modify the RISCV32 name system accoriding the AARCH64.
Improve the thread code in thread.hpp and thread.inline.hpp according the
http://hg.openjdk.java.net/jdk/jdk/rev/c153b4c52e39 .
After this commit, the make of slowdebug version can compiled success, but when run the
java -version, it still have the thread error. The error is same as the 'make images' of
slowdebug compiling.