Fix the data of lui/addi
When a 32 bit data used in lui-addi, we need to devision it to two numbers, one is
20bit and the other is 12bit, we will add them into one number in the future. So
we need to make sure they will get the same nubmer compared to the original number,
this need us to add a 0X800 into 20bit number or make 'upper = tagert -lower'.

This patch dealed with the lui-addi, also according the auipc-addi code.