change the relationship between j_rargx and c_rargx
The int reg don't have the reg like R1_H, so the next of R16 must R17. The j_rarg7 is c_rarg0,
it is the x10, but the j_rarg6 is c_rarg7 which is x17. The list can't be right in the next(),
so change the java function registers.