improve the div system
The code of div system need to improve, then we need to build a new div system for rv32g.
In rv32g div system, the int data use the div/rem, the long data need to deal with more
calcute insts.

This patch will clean the code, and delete some code we will not need. This patch will not
change the result of any type div.